feat(): move createProviderClass and unwrapProvider to general
