+++ packages/utils/src/general/index.ts
@@ -1,4 +1,6 @@
+export * from "./createProviderClass.js";
 export * from "./deepFreeze.js";
 export * from "./hasOwnProperty.js";
 export * from "./isObject.js";
 export * from "./JsonStorage.js";
+export * from "./unwrapProvider.js";
