+++ packages/utils/src/storyboard/index.ts
@@ -1,5 +1,10 @@
-export * from "./createProviderClass.js";
 export * from "./expressions/index.js";
 export * from "./functions/index.js";
 export * from "./permissions.js";
-export * from "./unwrapProvider.js";
+import { createProviderClass, unwrapProvider } from "@next-core/utils/general";
+export {
+  /** @deprecated import it from "@next-core/utils/general" instead */
+  createProviderClass,
+  /** @deprecated import it from "@next-core/utils/general" instead */
+  unwrapProvider,
+};
